Simplify EBNF grammar in preparation for `...` spread operator.
Note: for completeness of vision, the EBNF grammar shows where the `...`
spread operator will go, but `...` has not yet been enabled in the
parser. When we're ready, parsing `...` will be a one-hunk change.

More importantly, the KeyPath grammar rule has changed:
```diff
-KeyPath ::= Key PathStep+
+KeyPath ::= Key PathStep*
 PathStep ::= "." Key | "->" Identifier MethodArgs?
```
This means a KeyPath can now be a single key, like `b` and `c` in
`a->add(b, c)`, when there is no field/path ambiguity. When there *is*
ambiguity (within a SubSelection), `$.` can be used to anonymize a `key`
path, as in `$.key`, forcing a `PathSelection` interpretation. This
relaxation of rules around field/path ambiguity should prevent a great
deal of developer surprise by reducing the need for `$.` prefixes.

This simplification takes advantage of the recent unification of
`NamedSelection` enum variants into a single struct, because without
that refactoring we'd have to support the `...` operator separately for
both `NamedSelection::Field` and `NamedSelection::Path`, whereas now
there's only one AST structure where it can appear (see the `prefix:
NamingPrefix` field of the `NamedSelection` struct).