Future-proof grammar/parser to prevent multiple `?`s.
Eventually we may also want infix operators like `??` and `?!`, and it
should be possible to parse `a??b` or `a? ?? b?` without ambiguity.

This commit protects us from breaking changes by ensuring no more than
one consecutive `?` is ever considered to be a `PathList::Question`, and
by refusing to parse a single `?` if the text starts with `??` or `?!`.